refactor: introduce Poller.all_events_in_status?
